last changes 😃

+++ contracts/IFaucet.sol
@@ -2,6 +2,7 @@
 pragma solidity >=0.4.22 <0.9.0;
 
 interface IFacuet {
-    function  addFunds() external payable;
-    function withdraw(uint withdrawAmount) external;
+    function addFunds() external payable;
+
+    function withdraw(uint256 withdrawAmount) external;
 }
